requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.12
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
